Add files via upload
The bang ! file is like a README for each BGB work-in-progress directory, since the bang file name sorts to the top. Most of my SAILDART work assumes the existence of a large 2 Terabyte GNU/Linux file system that has multiple versions of the full DART tapes as well as a mysql data base with a lot of miscellaneous tables concerning SAIL software and people.